booking-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
mysql-connector-python
python-multipart
cryptography
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
jinja2